+++ Projects/Addons/Sources/Models/AddonModel.swift
@@ -4,23 +4,23 @@ import hGraphQL
 public struct AddonModel: Identifiable, Equatable, Hashable {
     public let id: String
     let title: String
-    let tag: String
     let subTitle: String?
+    let tag: String
     let informationText: String
     let options: [AddonOptionModel]
 
     public init(
         id: String,
         title: String,
-        tag: String,
         subTitle: String?,
+        tag: String,
         informationText: String,
         options: [AddonOptionModel]
     ) {
         self.id = id
         self.title = title
-        self.tag = tag
         self.subTitle = subTitle
+        self.tag = tag
         self.informationText = informationText
         self.options = options
     }
