+++ Projects/Addons/Sources/Service/DemoImplementation/AddonsClientDemo.swift
@@ -8,8 +8,8 @@ public class AddonsClientDemo: AddonsClient {
             .init(
                 id: "Travel Plus",
                 title: "Travel Plus",
-                tag: "Popular",
                 subTitle: "Extended travel insurance with extra coverage for your travels",
+                tag: "Popular",
                 informationText: "Click to learn more about our extended travel coverage Reseskydd Plus",
                 options: [
                     .init(
