+++ Projects/Addons/Sources/Views/AddonCardView.swift
@@ -52,8 +52,8 @@ public struct AddonCardView: View {
             addon: .init(
                 id: "id",
                 title: "Travel Plus",
-                tag: "Popular",
                 subTitle: "Extended travel insurance with extra coverage for your travels",
+                tag: "Popular",
                 informationText: "",
                 options: []
             )
