+++ Projects/Addons/Sources/Views/ChangeAddonScreen.swift
@@ -5,6 +5,7 @@ import hCoreUI
 struct ChangeAddonScreen: View {
     @EnvironmentObject var changeAddonNavigationVm: ChangeAddonNavigationViewModel
     @ObservedObject var changeAddonVm: ChangeAddonViewModel
+    @SwiftUI.Environment(\.colorScheme) private var colorScheme
 
     public init(
         changeAddonVm: ChangeAddonViewModel
@@ -108,7 +109,7 @@ struct ChangeAddonScreen: View {
                 Spacer()
 
                 hPill(
-                    text: "+ " + (addonOption.price?.formattedAmountPerMonth ?? ""),
+                    text: "+ " + (changeAddonVm.selectedSubOption?.price.formattedAmountPerMonth ?? ""),
                     color: .grey(translucent: true),
                     colorLevel: .one
                 )
@@ -120,8 +121,6 @@ struct ChangeAddonScreen: View {
                     .padding(.top, .padding8)
             }
 
-            let colorScheme: ColorScheme =
-                UITraitCollection.current.userInterfaceStyle == .light ? .light : .dark
             DropdownView(
                 value: String(changeAddonVm.selectedSubOption?.title ?? ""),
                 placeHolder: L10n.addonFlowSelectDaysPlaceholder
@@ -131,62 +130,6 @@ struct ChangeAddonScreen: View {
             .padding(.top, .padding16)
             .hBackgroundOption(option: (colorScheme == .light) ? [.negative] : [.secondary])
             .hSectionWithoutHorizontalPadding
-        }
-    }
-}
-
-@MainActor
-public class ChangeAddonViewModel: ObservableObject {
-    @Inject private var addonService: AddonsClient
-    @Published var viewState: ProcessingState = .loading
-    @Published var selectedSubOption: AddonSubOptionModel?
-    @Published var addonOptions: [AddonOptionModel]?
-    @Published var contractInformation: AddonContract?
-    @Published var informationText: String?
-
-    init(contractId: String) {
-        Task {
-            await getAddons()
-            await getContractInformation(contractId: contractId)
-
-            self._selectedSubOption = Published(
-                initialValue: addonOptions?.first?.subOptions.first
-            )
-        }
-    }
-
-    func getAddons() async {
-        withAnimation {
-            self.viewState = .loading
-        }
-
-        do {
-            let data = try await addonService.getAddon()
-
-            withAnimation {
-                self.addonOptions = data.options
-                self.informationText = data.informationText
-                self.viewState = .success
-            }
-        } catch let exception {
-            self.viewState = .error(errorMessage: exception.localizedDescription)
-        }
-    }
-
-    func getContractInformation(contractId: String) async {
-        withAnimation {
-            self.viewState = .loading
-        }
-
-        do {
-            let data = try await addonService.getContract(contractId: contractId)
-
-            withAnimation {
-                self.contractInformation = data
-                self.viewState = .success
-            }
-        } catch let exception {
-            self.viewState = .error(errorMessage: exception.localizedDescription)
         }
     }
 }
