+++ Projects/TravelCertificate/Sources/Views/ListScreen.swift
@@ -61,8 +61,8 @@ public struct ListScreen: View {
                             addon: .init(
                                 id: "id",
                                 title: "Travel Plus",
-                                tag: "Popular",
                                 subTitle: "Extended travel insurance with extra coverage for your travels",
+                                tag: "Popular",
                                 informationText: "",
                                 options: []
                             )
